Support Slim 4.11 for v1.1 release.  Closes #1.
Slim 4.11 adds PHP 7.4 types to the class we need to modify, further
requiring us to raise our minimum PHP version.  Previously, our
changelog stated we would do a major release in this scenario.

However, after undoubtedly overthinking it, I believe that users would
be best served by releasing two updates: a 1.0.x patch release that
prevents that line from co-installing with Slim 4.11, and a 1.1 minor
release that requires 4.11.

This means that "^1" will do the right thing, depending on how the rest
of the user's project chooses Slim.  There is no API breakage in our own
package, which is the primary concern of SemVer.