Add Containerfiles
With OpenSSL 3.0 on my host now, it was a much larger effort to get PHP
7.x built: I needed a private OpenSSL 1.1.1q installation.  I am not
looking forward to more libraries, or even whole toolchains, as the past
drifts further away.

Therefore, the testing paradigm changes from "code running on the host"
to "code running in a container."  The container uses _whatever is
checked out_ on the host, but runs it in one of the official PHP
container images.

A container can become frozen, locked to its contemporary environment...
but doing so gives us an extremely clear window into the past.